[improve] job submission parameter optimization
